# Conflicts:
#	project/src/main/java/model/Database.java

+++ project/src/main/java/model/NearbyVehicle.java
@@ -2,15 +2,15 @@ package model;
 
 public class NearbyVehicle extends Vehicle {
 
-    private final int distance;
+    private final double distance;
 
     protected NearbyVehicle(String registration, String make, String model, int year, String colour, Position position,
-	    int distance) {
+	    double distance) {
 	super(registration, make, model, year, colour, position);
 	this.distance = distance;
     }
 
-    public int getDistance() {
+    public double getDistance() {
 	return this.distance;
     }
